// src/app/modules/relationship-management/models/response/RemDistributionModel.ts
export interface RemDistributionModel {
	remDistributionId: number;
	mlabPlayerId: number;
	playerId: string;
	vipLevel: string;
	username: string;
	playerStatus: string;
	brand: string;
	currency: string;
	remProfileId: number;
	remProfileName: string;
	agentName: string;
	pseudoName: string;
	assignStatus: string;
	distributionDate: string;
	createdBy: string;
	lastAssigmentDate: string;
	assignedBy: string;
	prevReMProfileName: string;
	hasIntegration: number;
}